refactor(foundation): move HttpExchange to foundation + migrate recorder
Phase 6 / A3 follow-up: HttpExchange (used by mockforge-recorder for OpenAPI
generation) is a pure-data struct with only chrono + std deps. Moved to
mockforge-foundation::intelligent_behavior alongside Persona/Request/Response.

Migrated:
- mockforge-core/src/intelligent_behavior/openapi_generator.rs: re-exports
  HttpExchange from foundation
- mockforge-recorder/src/openapi_export.rs: imports HttpExchange from foundation
- mockforge-recorder/Cargo.toml: added mockforge-foundation dep
