+++ crates/mockforge-recorder/Cargo.toml
@@ -56,6 +56,7 @@ reqwest = { workspace = true }
 
 # Core types for OpenAPI generation
 mockforge-core = { version = "0.3.70", path = "../mockforge-core" }
+mockforge-foundation = { path = "../mockforge-foundation" }
 
 [dev-dependencies]
 tokio = { version = "1.48", features = ["macros", "rt-multi-thread"] }
